Rename `byte` to `Byte`

+++ tests/core/pyspec/eth_consensus_specs/utils/ssz/ssz_typing.py
@@ -2,7 +2,7 @@
 
 from remerkleable.basic import (
     boolean as Boolean,
-    byte,
+    byte as Byte,
     uint,
     uint8,
     uint16,
